// src/types/enums.ts
export enum CellEnum {
  Text = 'text',
  Currency = 'currency',
  Category = 'category',
  Availability = 'availability'
}

export enum StorageIconState {
  Empty,
  Low,
  Ok,
  Full
}

export enum IconEnum {
  Search = 'search',
  Storage = 'storage'
}

export enum CategoryColor {
  Red = '#E4ACB2',
  Orange = '#EABCA8',
  Yellow = '#FAEDCD',
  Green = '#CCD5AE',
  Blue = '#99BAB9'
}

export enum CategoryColorNames {
  Red = 'red',
  Orange = 'orange',
  Yellow = 'yellow',
  Green = 'green',
  Blue = 'blue'
}